test(api): address review — relocate allauth tests, trim redundant cases
Review of the prior commit surfaced three actions:

1. Move the allauth-headless tests (signup/verify, login/logout, password
   reset, registration gating) out of authentication/views_test.py into a new
   authentication/allauth_test.py. These exercise /_allauth/ (not v0/DRF) and
   would have been silently deleted with views_test.py during v0 removal.
   views_test.py is now purely v0.

2. Delete test_activation_already_active_user_is_idempotent — the activate()
   handler is branch-free, so this traversed the same lines as the two other
   activation tests; two of its assertions were tautological (asserted state
   the setup already established).

3. Drop test_me_patch_requires_auth (weakest of the auth-wiring guards;
   test_delete_requires_auth keeps the destructive route covered) and refresh
   a stale comment in main/api.py now that the ValidationError body shape has a
   dedicated test.
